Add timeout to deal with fake ultrasonic sensors
The HC-SR04 datasheet says that the senors will send at most a 38ms long
echo pulse, but this is not true of the chinese clones I have available.
They will send an indefinitely long echo pulse if they don't recieve an echo.
This leads to them locking up until one of the other sensors fires a pulse
that they can hear.

So, after 30ms, the sensor's power is cut for 100us. On some certain fakes
this leads to ECHO going low and staying low (on others, ECHO is pulled high
for ~60ms after this procedure).

I've found three sensors that work with this code and installed them.